updated to be compatible with v5.x of react-native-svg-charts.
