+++ package.json
@@ -29,11 +29,11 @@
     "symbol-observable": "^1.2.0"
   },
   "devDependencies": {
-    "babel-jest": "^23.6.0",
-    "metro-react-native-babel-preset": "^0.45.0",
+    "babel-jest": "^24.0.0",
     "cryptiles": "^4.1.3",
-    "jest": "^23.6.0",
+    "jest": "^24.0.0",
     "merge": ">=1.2.1",
+    "metro-react-native-babel-preset": "^0.45.0",
     "react-test-renderer": "16.7.0"
   },
   "jest": {
